implement Inheritance: Phone, Email and Address inherit newly added Contact class
